BS_DRF/requirements.txt
amqp==1.4.9
anyjson==0.3.3
asgiref==3.5.1
backports.zoneinfo==0.2.1
billiard==3.3.0.23
Django==4.0.4
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
kombu==3.0.37
psycopg2==2.9.3
PyJWT==2.3.0
pytz==2022.1
schedule==1.1.0
sqlparse==0.4.2
